[MOD] Put build_filter_case_ioc_query in business/iocs.py
